requirements.txt
streamlit
pandas
gspread
oauth2client
feedparser
google-generativeai
beautifulsoup4
requests